Feat: 게시판에 유저정보 활용 추가

+++ src/main/java/com/webtoonchat/toonchat/controller/board/BoardApiController.java
@@ -1,7 +1,9 @@
 package com.webtoonchat.toonchat.controller.board;
 
+import java.nio.charset.StandardCharsets;
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,24 +18,38 @@ import com.webtoonchat.toonchat.domain.board.Board;
 import com.webtoonchat.toonchat.dto.board.AddBoardRequest;
 import com.webtoonchat.toonchat.dto.board.BoardResponse;
 import com.webtoonchat.toonchat.dto.board.UpdateBoardRequest;
+import com.webtoonchat.toonchat.security.jwt.util.JwtTokenizer;
 import com.webtoonchat.toonchat.service.board.BoardService;
 
+import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+
+@Slf4j
 @RequiredArgsConstructor
 @RestController
 public class BoardApiController {
 	private final BoardService boardService;
+	private final JwtTokenizer jwtTokenizer;
+
+	@Value("${jwt.secretKey}")
+	private String secretKey;
 
 	@PostMapping("/api/boards/{bgno}")
-	public ResponseEntity<Board> addBoard(@PathVariable String bgno, @RequestBody AddBoardRequest request) {
-		Board savedArticle = boardService.save(request, bgno);
+	public ResponseEntity<Board> addBoard(
+			@PathVariable String bgno, @RequestBody AddBoardRequest request, HttpServletRequest httpServletRequest) {
+		Claims claims = extracted(httpServletRequest);
+		Integer userid = (Integer) claims.get("userId");
+		Board savedArticle = boardService.save(request, bgno, userid);
 		return ResponseEntity.status(HttpStatus.CREATED)
 				.body(savedArticle);
 	}
 
 	@GetMapping("/api/boards/{bgno}")
-	public ResponseEntity<List<BoardResponse>> findAllBoards(@PathVariable String bgno) {
+	public ResponseEntity<List<BoardResponse>> findAllBoards(
+			@PathVariable String bgno, HttpServletRequest httpServletRequest) {
 		List<BoardResponse> articles = boardService.findAllByBgno(bgno)
 				.stream()
 				.map(BoardResponse::new)
@@ -44,7 +60,8 @@ public class BoardApiController {
 	}
 
 	@GetMapping("/api/boards/{bgno}/{id}")
-	public ResponseEntity<BoardResponse> findBoard(@PathVariable String bgno, @PathVariable long id) {
+	public ResponseEntity<BoardResponse> findBoard(
+			@PathVariable String bgno, @PathVariable long id, HttpServletRequest httpServletRequest) {
 		Board article = boardService.findById(id);
 
 		return ResponseEntity.ok()
@@ -52,7 +69,14 @@ public class BoardApiController {
 	}
 
 	@DeleteMapping("/api/boards/{bgno}/{id}")
-	public ResponseEntity<Void> deleteBoard(@PathVariable String bgno, @PathVariable long id) {
+	public ResponseEntity<Void> deleteBoard(
+			@PathVariable String bgno, @PathVariable long id, HttpServletRequest httpServletRequest) {
+		Claims claims = extracted(httpServletRequest);
+		if (boardService.findById(id).getWriterId() != (int)claims.get("userId")) {
+			log.error("작성자가 아닌 글 삭제 요청");
+			return ResponseEntity.badRequest()
+					.build();
+		}
 		boardService.delete(id);
 		return ResponseEntity.ok()
 				.build();
@@ -62,11 +86,25 @@ public class BoardApiController {
 	public ResponseEntity<Board> updateBoard(
 			@PathVariable String bgno,
 			@PathVariable long id,
-			@RequestBody UpdateBoardRequest request
+			@RequestBody UpdateBoardRequest request,
+			HttpServletRequest httpServletRequest
 	) {
+		Claims claims = extracted(httpServletRequest);
+		if (boardService.findById(id).getWriterId() != (int)claims.get("userId")) {
+			log.error("작성자가 아닌 글 수정 요청");
+			return ResponseEntity.badRequest()
+					.build();
+		}
 		Board updateArticle = boardService.update(id, request);
 
 		return ResponseEntity.ok()
 				.body(updateArticle);
+	}
+
+	private Claims extracted(HttpServletRequest httpServletRequest) {
+		String authorization = httpServletRequest.getHeader("Authorization");
+		String[] arr = authorization.split(" ");
+		log.info(arr[1]);
+		return jwtTokenizer.parseToken(arr[1], secretKey.getBytes(StandardCharsets.UTF_8));
 	}
 }
